+++ packages/apollo-boost/CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ### vNext
 
-- Allow instantiating the client with `headers` and `credentials`.
+- Allow `headers` and `credentials` to be passed in as configuration
+  parameters to the `apollo-boost` `ApolloClient` constructor.
+  [PR #3098](https://github.com/apollographql/apollo-client/pull/3098)
 
 ### 0.1.7