+++ packages/apollo-boost/CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+### vNext
+
+- Allow `headers` and `credentials` to be passed in as configuration
+  parameters to the `apollo-boost` `ApolloClient` constructor.
+  [PR #3098](https://github.com/apollographql/apollo-client/pull/3098)
+
 ### 0.1.7
 
 - No public facing functionality changes.