+++ packages/apollo-boost/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### vNext
 
+- The `apollo-boost` `ApolloClient` constructor warns about unsupported options.
+  [PR #3551](https://github.com/apollographql/apollo-client/pull/3551)
 - Allow `headers` and `credentials` to be passed in as configuration
   parameters to the `apollo-boost` `ApolloClient` constructor.
   [PR #3098](https://github.com/apollographql/apollo-client/pull/3098)